Code reorganization of the ffiwrappers (in ffiwrappers)
