+++ ffiwrappers/include/wrappingcore.h
@@ -1,0 +1,6 @@
+
+size_t bk_getfn(const char* fname);
+
+// Not a prototype, so warning will be generated. TODO: switch off
+// warning for this instance only?
+#define BKFNPY(F)  (* ( void (*)() )(bk_getfn( #F )))
